+++ src/dftbp/dftbplus/parser.F90
@@ -1141,6 +1141,7 @@ contains
     type(TStatus), intent(inout) :: errStatus
 
     type(string) :: buffer
+    type(fnode), pointer :: child
 
     call getNodeName(node, buffer)
     select case (char(buffer))
@@ -1161,6 +1162,15 @@ contains
     case default
       call detailedError(node, "Invalid Hamiltonian")
     end select
+
+  #:if WITH_API
+    call getChild(node, "ASI", child, requested=.false.)
+    if (associated(child)) then
+      ctrl%isASICallbackEnabled = .true.
+    else
+      ctrl%isASICallbackEnabled = .false.
+    end if
+  #:endif
 
   end subroutine readHamiltonian